requirements.txt
fastapi == 0.100.1
uvicorn[standard]
python-multipart
PyPDF2
langchain==0.1.6
openai==0.27.8
tiktoken
faiss-cpu
textract
openpyxl
pandas
langchain-community==0.0.19
langchain-core==0.1.23
torch
accelerate
transformers
sentence_transformers
streamlit
streamlit_chat
altair
huggingface-hub
google-generativeai
pillow
langchain-google-genai
python-dotenv
langchain-google-vertexai
ipython
duckduckgo-search
pypdf